requirements.txt
acryl-datahub==0.15.0.5
acryl-datahub-classify==0.0.11
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
altair==4.2.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
avro==1.11.3
avro-gen3==0.7.16
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blis==0.7.11
cached-property==2.0.1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-default-group==1.2.4
click-spinner==0.1.10
cloudpathlib==0.21.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
datahub==0.999.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
docker==7.1.0
dotenv==0.9.9
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
expandvars==0.12.0
fastjsonschema==2.21.1
fqdn==1.5.1
frozenlist==1.5.0
great-expectations==0.15.50
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.10
ijson==3.3.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipaddress==1.0.23
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
langcodes==3.5.0
language_data==1.3.0
makefun==1.15.6
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
mixpanel==4.10.1
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patchy==2.8.0
pexpect==4.9.0
phonenumbers==8.13.0
platformdirs==4.3.6
preshed==3.0.9
progressbar2==4.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycountry==24.6.1
pycparser==2.22
pydantic==1.10.21
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-stdnum==1.20
python-utils==3.9.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rstr==3.2.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.12
schwifty==2024.6.1
scipy==1.13.1
Send2Trash==1.8.3
sentry-sdk==2.22.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
sqlglot==25.32.1
sqlglotrs==0.3.0
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
tabulate==0.9.0
terminado==0.18.1
thinc==8.2.5
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uri-template==1.3.0
urllib3==1.26.20
vininfo==1.8.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
